rose/ns/requirements.txt
dnspython==1.15.0
jinja2==2.10
